temp/requirements.txt
# Core
python-dotenv==1.0.0

# LangChain and LangGraph - compatible versions
langchain-core==0.1.23
langchain-google-genai==0.0.6
langgraph==0.0.20

# Google Gemini
google-generativeai==0.3.2

# Pydantic v1 (for compatibility with langgraph)
pydantic==1.10.13

# Utilities
typing-extensions==4.9.0

